GG-39017 Extract handling of shutdown policy to a separate class
